use chisme unit tests
