+++ charms/tensorboards-web-app/requirements.txt
@@ -1,6 +1,6 @@
 # Copyright 2021 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-ops==1.2.0
+ops >= 1.2.0
 oci-image<1.1.0
 serialized-data-interface<0.4.0
