+++ charms/tensorboards-web-app/tests/unit/test_charm.py
@@ -3,7 +3,10 @@
 
 import pytest
 
-from ops.model import ActiveStatus, BlockedStatus, WaitingStatus
+from charmed_kubeflow_chisme.testing import test_leadership_events as leadership_events
+from charmed_kubeflow_chisme.testing import test_missing_image as missing_image
+from charmed_kubeflow_chisme.testing import test_missing_relation as missing_relation
+from ops.model import ActiveStatus, BlockedStatus
 from ops.testing import Harness
 import yaml
 
@@ -15,15 +18,31 @@ def harness():
     return Harness(Operator)
 
 
-def test_not_leader(harness):
-    harness.begin_with_initial_hooks()
-    assert isinstance(harness.charm.model.unit.status, WaitingStatus)
+def test_leadership_events(harness):
+    leadership_events(harness)
 
 
 def test_missing_image(harness):
+    missing_image(harness, BlockedStatus)
+
+
+def test_missing_relation(harness):
+    missing_relation(harness, ActiveStatus)
+
+
+def test_no_relation(harness):
     harness.set_leader(True)
+    harness.add_oci_resource(
+        "oci-image",
+        {
+            "registrypath": "ci-test",
+            "username": "",
+            "password": "",
+        },
+    )
     harness.begin_with_initial_hooks()
-    assert isinstance(harness.charm.model.unit.status, BlockedStatus)
+
+    assert harness.charm.model.unit.status == ActiveStatus()
 
 
 def test_main(harness):
